Restructuring Refactors
