+++ app/src/main/java/com/bintina/goouttolunchmvvm/user/login/OnLogInOnClickListener.kt
@@ -1,4 +1,4 @@
-package com.bintina.goouttolunchmvvm.login
+package com.bintina.goouttolunchmvvm.user.login
 
 interface OnLogInOnClickListener {
